requirements.txt
# Core financial data libraries
mftool==1.1.0
mstarpy>=0.0.5
jugaad-data>=0.25
morningstar-fund-scraper==0.1.0
nsescraper==0.0.8

# Data processing and validation
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# Utilities
python-dateutil>=2.8.2
requests>=2.31.0
pyyaml>=6.0
beautifulsoup4>=4.12.2

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging and configuration
python-dotenv>=1.0.0
